added time stamp

+++ Maintendance/attendanceentry.cpp
@@ -11,5 +11,5 @@ QString AttendanceEntry::_getDate()const
     return this->attDate;
 }
 
-AttendanceEntry::AttendanceEntry(QString d, QString date, bool present, int hrs)
+AttendanceEntry::AttendanceEntry(QString d, QString date, bool present, double hrs)
     : day(d), attDate(date), attendance(present), hours(hrs) {}
